Moving models surrounding data into their respective new folder. updating references
