+++ src/models/data_sizes/Word.ts
